Parsing code from 2005
This is committed because it was a useful exercise in creating a parser
in a test first fashion. It creates a restricted grammar and subset of
a language, which is precisely what I want to do here. It's relevant
because the grammar here is actually more complex, because it does
not have keywords per se which makes life somewhat harder than normal
programming languages.