Added microfont.odt file
This was used during the creation/design of the font included here.
In particular it's derived from the open source font "tiny" which
is available here:

    http://www.squaregear.net/fonts/tiny.shtml

In particular, this is a permissive BSD/MIT style license. In many
respects it's used as inspiration, or for lower case letters. However
it's not used directly, because it's not available directly in a form
that's usable.