First command line program for the device
This is the same program that I've used for testing the device to date,
but just built and installed using command line tools. Specifically,
you do this:

   make
   <press reset on device>
   make upload

This results in the code being uploaded to the device.

This code loops through all the pixels displaying them, and based on
button presses changes the behaviour.  While simplistic, it's useful
for testing the LEDS..