2Mhz base for timer, reorder behaviour in display handler
For some reason the manually created microbug is running at 2Mhz - or
the timers are based off that, based on a manual calculation. Changed
timers based on this.

Secondly, the LEDs used in the microbug have a faster response time,
so the code showed up ghosting if switches were toggled in a particular
order. They're now toggled in an order that prevents that.