Switch font repr to unsigned char* (halves size)
Changed the font representation to unsigned chars - this halves the
size of the data used. Next step is to shift this into progmem, and
perhaps use a nibble based representation, given we only use 4 bits
per row.

I suspect the shift into progmem will be the most beneficial change
to be fair.