Sync APIs between the two DAL cores
The two DAL cores are legitimate alternatives. One is powered and one
is low power. If they both provide the same external API to daluser,
it simplifies the implementation of having both dramatically, without
sacrficing brightness in tethered mode.

In particular, the ordering of functions has also been sync'd both in terms
of prototypes and implementation.

The exception is the tether DAL core still contains a function it shouldn't
related to displaying a 5x5 image array.